OpenML Collection

This is the class for collections (previously known as studies) served on https://openml.org/search?type=study&study_type=task&sort=tasks_included. It is used both for Run Collections and Task Collections. (Note that all Benchmark Suites on OpenML are also Collections). A Run Collection (main_entity_type = "run") contains runs, flows, datasets and tasks. A Task Collection (main_entity_type = "task") contains tasks and datasets.

mlr3 Intergration

Obtain a list of mlr3::Tasks using as_tasks().

Obtain a list of mlr3::Resamplings using as_resamplings().

Obtain a list of mlr3::Learners using as_learners() (if main_entity_type is "run").

Obtain a mlr3::BenchmarkResult using as_benchmark_result() (if main_entity_type is "run").

References

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi:10.1145/2641190.2641198 .

Public fields

id

(integer(1))
OpenML collection id.

cache_dir

(logical(1) | character(1))
Stores the location of the cache for objects retrieved from https://openml.org. If set to FALSE, caching is disabled.

The package qs is required for caching.

Active bindings

desc

(list())
Colllection description (meta information), downloaded and converted from the JSON API response.

name

(character(1))
The name of the collection.

tags

(character(n))
The tags of the OpenML collection.

main_entity_type

(character(1))
The main entity type of the collection (either "run" or "task").

flow_ids

(integer(n))
An vector containing the flow ids of the collection.

data_ids

(integer(n))
An vector containing the data ids of the collection.

run_ids

(integer(n))
An vector containing the run ids of the collection.

task_ids

(integer(n))
An vector containing the task ids of the collection.

runs

(data.table()) A data.table summarizing the runs included in the collection. Returns NULL for Task Collections.

flows

(data.table()) A data.table summarizing the flows included in the collection. Returns NULL for Task Collections.

data

(data.table()) A data.table summarizing the datasets included in the collection.

tasks

(data.table()) A data.table summarizing the tasks included in the collection.

Methods

Public methods

OMLCollection$new()

OMLCollection$print()

OMLCollection$clone()

Method new()

Creates a new object of class OMLCollection.

Usage

OMLCollection$new(id, cache = getOption("mlr3oml.cache", FALSE))

Arguments

id

(integer(1))
OpenML run id.

cache

(logical(1) | character(1))
See field cache for an explanation of possible values. Defaults to value of option "mlr3oml.cache", or FALSE if not set.

Method print()

Prints the object.

Usage

OMLCollection$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

OMLCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library("mlr3")
# OpenML Run Collection:
collection = OMLCollection$new(232L)
collection$tasks
#>    id          task     data                 task_type target  nrow ncol
#> 1:  3 <OMLTask[18]> kr-vs-kp Supervised Classification  class  3196   37
#> 2:  6 <OMLTask[18]>   letter Supervised Classification  class 20000   17
#>    missing numeric symbolic binary      data_split
#> 1:       0       0       37     35 crossvalidation
#> 2:       0      16        1      0 crossvalidation
collection$data
#>    id          data     name  nrow ncol missing numeric symbolic binary
#> 1:  3 <OMLData[19]> kr-vs-kp  3196   37       0       0       37     35
#> 2:  6 <OMLData[19]>   letter 20000   17       0      16        1      0
collection$flows
#>       id          flow
#> 1: 17369 <OMLFlow[11]>
#> 2: 17374 <OMLFlow[11]>
#>                                                                                                                                                                             name
#> 1: sklearn.model_selection._search_successive_halving.HalvingRandomSearchCV(estimator=sklearn.ensemble._hist_gradient_boosting.gradient_boosting.HistGradientBoostingClassifier)
#> 2:                       sklearn.model_selection._search.RandomizedSearchCV(estimator=sklearn.ensemble._hist_gradient_boosting.gradient_boosting.HistGradientBoostingClassifier)
collection$runs
#>          id          run                 task_type     data
#> 1: 10417460 <OMLRun[18]> Supervised Classification kr-vs-kp
#> 2: 10417461 <OMLRun[18]> Supervised Classification kr-vs-kp
#> 3: 10417462 <OMLRun[18]> Supervised Classification   letter
#> 4: 10417463 <OMLRun[18]> Supervised Classification   letter
#>                              flow      data_split
#> 1: sklearn.model_selection._[...] crossvalidation
#> 2: sklearn.model_selection._[...] crossvalidation
#> 3: sklearn.model_selection._[...] crossvalidation
#> 4: sklearn.model_selection._[...] crossvalidation

# mlr3 conversion:
tasks = as_tasks(collection)
resamplings = as_resamplings(collection)
# construct pseudo-learners as these are sklearn flows
learners = as_learners(collection, "classif")

# Although pseudo-learners are non-executable the runs can still be analyzed.
bmr = as_benchmark_result(collection)
bmr$score(msr("classif.ce"))
#>                                    uhash nr              task  task_id
#>  1: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  2: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  3: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  4: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  5: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  6: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  7: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  8: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#>  9: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#> 10: f4d22a15-5415-4003-80bd-f66dc8c820c5  1 <TaskClassif[50]> kr-vs-kp
#> 11: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 12: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 13: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 14: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 15: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 16: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 17: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 18: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 19: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 20: ff87ad92-fc81-429c-93b5-832be1a6086c  2 <TaskClassif[50]> kr-vs-kp
#> 21: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 22: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 23: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 24: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 25: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 26: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 27: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 28: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 29: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 30: 26b098d3-99a5-4fad-ae89-a3874492a935  3 <TaskClassif[50]>   letter
#> 31: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 32: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 33: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 34: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 35: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 36: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 37: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 38: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 39: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#> 40: ffe6113e-9ae1-495c-a09b-8104c1f1b724  4 <TaskClassif[50]>   letter
#>                                    uhash nr              task  task_id
#>                          learner learner_id         resampling resampling_id
#>  1: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  2: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  3: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  4: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  5: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  6: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  7: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  8: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#>  9: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 10: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 11: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 12: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 13: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 14: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 15: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 16: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 17: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 18: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 19: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 20: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 21: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 22: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 23: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 24: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 25: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 26: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 27: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 28: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 29: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 30: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCV[20]>            cv
#> 31: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 32: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 33: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 34: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 35: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 36: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 37: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 38: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 39: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#> 40: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCV[20]>            cv
#>                          learner learner_id         resampling resampling_id
#>     iteration              prediction  classif.ce
#>  1:         1 <PredictionClassif[20]> 0.003134796
#>  2:         2 <PredictionClassif[20]> 0.009375000
#>  3:         3 <PredictionClassif[20]> 0.000000000
#>  4:         4 <PredictionClassif[20]> 0.003125000
#>  5:         5 <PredictionClassif[20]> 0.006250000
#>  6:         6 <PredictionClassif[20]> 0.000000000
#>  7:         7 <PredictionClassif[20]> 0.003134796
#>  8:         8 <PredictionClassif[20]> 0.006269592
#>  9:         9 <PredictionClassif[20]> 0.003134796
#> 10:        10 <PredictionClassif[20]> 0.006269592
#> 11:         1 <PredictionClassif[20]> 0.006269592
#> 12:         2 <PredictionClassif[20]> 0.009375000
#> 13:         3 <PredictionClassif[20]> 0.003125000
#> 14:         4 <PredictionClassif[20]> 0.003125000
#> 15:         5 <PredictionClassif[20]> 0.012500000
#> 16:         6 <PredictionClassif[20]> 0.000000000
#> 17:         7 <PredictionClassif[20]> 0.003134796
#> 18:         8 <PredictionClassif[20]> 0.009404389
#> 19:         9 <PredictionClassif[20]> 0.006269592
#> 20:        10 <PredictionClassif[20]> 0.009404389
#> 21:         1 <PredictionClassif[20]> 0.047523762
#> 22:         2 <PredictionClassif[20]> 0.039500000
#> 23:         3 <PredictionClassif[20]> 0.043500000
#> 24:         4 <PredictionClassif[20]> 0.041000000
#> 25:         5 <PredictionClassif[20]> 0.030500000
#> 26:         6 <PredictionClassif[20]> 0.044000000
#> 27:         7 <PredictionClassif[20]> 0.045000000
#> 28:         8 <PredictionClassif[20]> 0.036000000
#> 29:         9 <PredictionClassif[20]> 0.039000000
#> 30:        10 <PredictionClassif[20]> 0.031000000
#> 31:         1 <PredictionClassif[20]> 0.047523762
#> 32:         2 <PredictionClassif[20]> 0.039500000
#> 33:         3 <PredictionClassif[20]> 0.043500000
#> 34:         4 <PredictionClassif[20]> 0.041000000
#> 35:         5 <PredictionClassif[20]> 0.030500000
#> 36:         6 <PredictionClassif[20]> 0.044000000
#> 37:         7 <PredictionClassif[20]> 0.045000000
#> 38:         8 <PredictionClassif[20]> 0.036000000
#> 39:         9 <PredictionClassif[20]> 0.039000000
#> 40:        10 <PredictionClassif[20]> 0.031000000
#>     iteration              prediction  classif.ce
# }