List Tasks from OpenML

This function allows to find tasks on https://openml.org/t using some simple filter criteria.

Note that only a subset of filters is exposed here. For a more feature-complete package, see OpenML.

list_oml_tasks(
  number_instances = NULL,
  number_features = NULL,
  number_classes = NULL,
  number_missing_values = NULL,
  tag = NULL,
  limit = 5000L,
  ...
)

Arguments

number_instances
(integer())
Filter for number of instances.

number_features
(integer())
Filter for number of features.

number_classes
(integer())
Filter for number of labels of the target (only classification tasks).

number_missing_values
(integer())
Filter for number of missing values.

tag
(character())
Filter for specific tag. You can provide multiple tags as character vector.

limit
(integer())
Limit the results to limit records. Default is 5000.

...
(any)
Additional filters as named arguments.

Value

(data.table()) of results.

Details

Filter values can be provided as single atomic values (typically integer or character). Provide a numeric vector of length 2 (c(l, u)) to find matches in the range \([l, u]\).

References

Casalicchio G, Bossek J, Lang M, Kirchhoff D, Kerschke P, Hofner B, Seibold H, Vanschoren J, Bischl B (2017). “OpenML: An R Package to Connect to the Machine Learning Platform OpenML.” Computational Statistics, 1--15. doi: 10.1007/s00180-017-0742-2 .

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi: 10.1145/2641190.2641198 .

Examples

# \donttest{
list_oml_tasks(number_instances = 150, number_features = c(1, 10))

#>      task_id                 task_type data_id            name status
#>   1:      59 Supervised Classification      61            iris active
#>   2:     118            Learning Curve      61            iris active
#>   3:     289 Supervised Classification      61            iris active
#>   4:    1758            Learning Curve      61            iris active
#>   5:    1823 Supervised Classification      61            iris active
#>  ---                                                                 
#> 171:  232824                Clustering   42098            iris active
#> 172:  232891                Clustering   42186 JuanFeldmanIris active
#> 173:  232938                Clustering   42261    iris-example active
#> 174:  233207            Learning Curve      61            iris active
#> 175:  233208            Learning Curve      61            iris active
#>      MajorityClassSize MaxNominalAttDistinctValues MinorityClassSize
#>   1:                50                           3                50
#>   2:                50                           3                50
#>   3:                50                           3                50
#>   4:                50                           3                50
#>   5:                50                           3                50
#>  ---                                                                
#> 171:                50                           3                50
#> 172:                50                           3                50
#> 173:                50                          NA                50
#> 174:                50                           3                50
#> 175:                50                           3                50
#>      NumberOfClasses NumberOfFeatures NumberOfInstances
#>   1:               3                5               150
#>   2:               3                5               150
#>   3:               3                5               150
#>   4:               3                5               150
#>   5:               3                5               150
#>  ---                                                   
#> 171:               3                5               150
#> 172:               3                5               150
#> 173:               3                5               150
#> 174:               3                5               150
#> 175:               3                5               150
#>      NumberOfInstancesWithMissingValues NumberOfMissingValues
#>   1:                                  0                     0
#>   2:                                  0                     0
#>   3:                                  0                     0
#>   4:                                  0                     0
#>   5:                                  0                     0
#>  ---                                                         
#> 171:                                  0                     0
#> 172:                                  0                     0
#> 173:                                  0                     0
#> 174:                                  0                     0
#> 175:                                  0                     0
#>      NumberOfNumericFeatures NumberOfSymbolicFeatures
#>   1:                       4                        1
#>   2:                       4                        1
#>   3:                       4                        1
#>   4:                       4                        1
#>   5:                       4                        1
#>  ---                                                 
#> 171:                       4                        1
#> 172:                       4                        1
#> 173:                       4                        1
#> 174:                       4                        1
#> 175:                       4                        1

# }