OpenML Collection

This is the class for collections (previously known as studies) served on https://openml.org. A collection can either be a task collection or run collection. This object can also be constructed using the sugar function oml_collection().

Run Collection

A run collection contains runs, flows, datasets and tasks. The primary object are the runs (main_entity_type is "run"). The the flows, datasets and tasks are those used in the runs.

Task Collection A task collection (main_entity_type = "task") contains tasks and datasets. The primary object are the tasks (main_entity_type is "task"). The datasets are those used in the tasks.

Note: All Benchmark Suites on OpenML are also collections.

Caching

The OpenML collection itself cannot be not cached, this is because it can be modified in-place on the server, e.g. by adding or removing tasks or runs. The construction argument cache therefore only controls wether caching is applied to the OpenML objects that are contained in the collection.

mlr3 Intergration

Obtain a list of mlr3::Tasks using mlr3::as_tasks.

Obtain a list of mlr3::Resamplings using mlr3::as_resamplings.

Obtain a list of mlr3::Learners using mlr3::as_learners (if main_entity_type is "run").

Obtain a mlr3::BenchmarkResult using mlr3::as_benchmark_result (if main_entity_type is "run").

References

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi:10.1145/2641190.2641198 .

Super class

mlr3oml::OMLObject -> OMLCollection

Active bindings

desc

(list())
Colllection description (meta information), downloaded and converted from the JSON API response.

parquet

(logical(1))
Whether to use parquet.

main_entity_type

(character(n))
The main entity type, either "run" or "task".

flow_ids

(integer(n))
An vector containing the flow ids of the collection.

data_ids

(integer(n))
An vector containing the data ids of the collection.

run_ids

(integer(n))
An vector containing the run ids of the collection.

task_ids

(integer(n))
An vector containing the task ids of the collection.

runs

(data.table()) A data.table summarizing the runs included in the collection. Returns NULL for Task Collections.

flows

(data.table()) A data.table summarizing the flows included in the collection. Returns NULL for Task Collections.

data

(data.table()) A data.table summarizing the datasets included in the collection.

tasks

(data.table()) A data.table summarizing the tasks included in the collection.

Methods

Public methods

OMLCollection$new()

OMLCollection$print()

OMLCollection$clone()

Inherited methods

mlr3oml::OMLObject$help()

Method new()

Creates a new instance of this R6 class.

Usage

OMLCollection$new(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", FALSE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)

Arguments

id

(integer(1))
OpenML id for the object.

cache

(logical(1) | character(1))
See field cache for an explanation of possible values. Defaults to value of option "mlr3oml.cache", or FALSE if not set. The collection itself is not cached, this is because it can be modified in-place on OpenML, e.g. by adding or removing tasks or runs. This parameter therefore only controls whether the contained elements are cached when loaded, e.g. when accessing the included tasks.

parquet

(logical(1))
Whether to use parquet instead of arff. If parquet is not available, it will fall back to arff.

test_server

(character(1))
Whether to use the OpenML test server (https://test.openml.org/) or public server (https://www.openml.org/). Defaults to value of option "mlr3oml.test_server", or FALSE if not set.

Method print()

Prints the object.

Usage

OMLCollection$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

OMLCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library("mlr3")
# OpenML Run collection:
run_collection = OMLCollection$new(id = 232)
# using sugar
run_collection = oml_collection(id = 232)
run_collection$main_entity_type
#> [1] "run"
run_collection$tasks
#>    id          task     data                 task_type target  nrow ncol
#> 1:  3 <OMLTask[25]> kr-vs-kp Supervised Classification  class  3196   37
#> 2:  6 <OMLTask[25]>   letter Supervised Classification  class 20000   17
#>    missing numeric symbolic binary     task_splits
#> 1:       0       0       37     35 crossvalidation
#> 2:       0      16        1      0 crossvalidation
run_collection$data
#>    id          data     name  nrow ncol missing numeric symbolic binary
#> 1:  3 <OMLData[25]> kr-vs-kp  3196   37       0       0       37     35
#> 2:  6 <OMLData[25]>   letter 20000   17       0      16        1      0
run_collection$flows
#>       id          flow                           name
#> 1: 17369 <OMLFlow[16]> sklearn.model_selection._[...]
#> 2: 17374 <OMLFlow[16]> sklearn.model_selection._[...]
run_collection$runs
#>          id          run                 task_type     data
#> 1: 10417460 <OMLRun[25]> Supervised Classification kr-vs-kp
#> 2: 10417461 <OMLRun[25]> Supervised Classification kr-vs-kp
#> 3: 10417462 <OMLRun[25]> Supervised Classification   letter
#> 4: 10417463 <OMLRun[25]> Supervised Classification   letter
#>                              flow     task_splits
#> 1: sklearn.model_selection._[...] crossvalidation
#> 2: sklearn.model_selection._[...] crossvalidation
#> 3: sklearn.model_selection._[...] crossvalidation
#> 4: sklearn.model_selection._[...] crossvalidation

# mlr3 conversion:
tasks = as_tasks(run_collection)
resamplings = as_resamplings(run_collection)
learners = as_learners(run_collection, "classif")

bmr = as_benchmark_result(run_collection)
bmr$score(msr("classif.ce"))
#>                                    uhash nr              task  task_id
#>  1: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  2: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  3: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  4: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  5: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  6: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  7: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  8: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#>  9: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#> 10: 22b51e30-bf1a-4465-b835-56fad9debcb6  1 <TaskClassif[50]> kr-vs-kp
#> 11: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 12: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 13: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 14: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 15: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 16: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 17: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 18: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 19: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 20: dffb1190-d308-4d5a-a4ee-d399798648d7  2 <TaskClassif[50]> kr-vs-kp
#> 21: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 22: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 23: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 24: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 25: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 26: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 27: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 28: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 29: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 30: 51645b74-b84d-413a-865c-d35f95ac0c43  3 <TaskClassif[50]>   letter
#> 31: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 32: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 33: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 34: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 35: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 36: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 37: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 38: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 39: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#> 40: bf33c9a6-28a3-4c5b-964f-8a5abd6ffb93  4 <TaskClassif[50]>   letter
#>                                    uhash nr              task  task_id
#>                          learner learner_id             resampling
#>  1: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  2: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  3: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  4: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  5: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  6: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  7: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  8: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  9: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 10: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 11: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 12: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 13: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 14: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 15: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 16: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 17: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 18: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 19: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 20: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 21: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 22: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 23: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 24: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 25: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 26: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 27: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 28: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 29: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 30: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 31: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 32: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 33: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 34: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 35: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 36: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 37: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 38: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 39: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 40: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#>                          learner learner_id             resampling
#>     resampling_id iteration              prediction  classif.ce
#>  1:        custom         1 <PredictionClassif[20]> 0.003134796
#>  2:        custom         2 <PredictionClassif[20]> 0.009375000
#>  3:        custom         3 <PredictionClassif[20]> 0.000000000
#>  4:        custom         4 <PredictionClassif[20]> 0.003125000
#>  5:        custom         5 <PredictionClassif[20]> 0.006250000
#>  6:        custom         6 <PredictionClassif[20]> 0.000000000
#>  7:        custom         7 <PredictionClassif[20]> 0.003134796
#>  8:        custom         8 <PredictionClassif[20]> 0.006269592
#>  9:        custom         9 <PredictionClassif[20]> 0.003134796
#> 10:        custom        10 <PredictionClassif[20]> 0.006269592
#> 11:        custom         1 <PredictionClassif[20]> 0.006269592
#> 12:        custom         2 <PredictionClassif[20]> 0.009375000
#> 13:        custom         3 <PredictionClassif[20]> 0.003125000
#> 14:        custom         4 <PredictionClassif[20]> 0.003125000
#> 15:        custom         5 <PredictionClassif[20]> 0.012500000
#> 16:        custom         6 <PredictionClassif[20]> 0.000000000
#> 17:        custom         7 <PredictionClassif[20]> 0.003134796
#> 18:        custom         8 <PredictionClassif[20]> 0.009404389
#> 19:        custom         9 <PredictionClassif[20]> 0.006269592
#> 20:        custom        10 <PredictionClassif[20]> 0.009404389
#> 21:        custom         1 <PredictionClassif[20]> 0.047523762
#> 22:        custom         2 <PredictionClassif[20]> 0.039500000
#> 23:        custom         3 <PredictionClassif[20]> 0.043500000
#> 24:        custom         4 <PredictionClassif[20]> 0.041000000
#> 25:        custom         5 <PredictionClassif[20]> 0.030500000
#> 26:        custom         6 <PredictionClassif[20]> 0.044000000
#> 27:        custom         7 <PredictionClassif[20]> 0.045000000
#> 28:        custom         8 <PredictionClassif[20]> 0.036000000
#> 29:        custom         9 <PredictionClassif[20]> 0.039000000
#> 30:        custom        10 <PredictionClassif[20]> 0.031000000
#> 31:        custom         1 <PredictionClassif[20]> 0.047523762
#> 32:        custom         2 <PredictionClassif[20]> 0.039500000
#> 33:        custom         3 <PredictionClassif[20]> 0.043500000
#> 34:        custom         4 <PredictionClassif[20]> 0.041000000
#> 35:        custom         5 <PredictionClassif[20]> 0.030500000
#> 36:        custom         6 <PredictionClassif[20]> 0.044000000
#> 37:        custom         7 <PredictionClassif[20]> 0.045000000
#> 38:        custom         8 <PredictionClassif[20]> 0.036000000
#> 39:        custom         9 <PredictionClassif[20]> 0.039000000
#> 40:        custom        10 <PredictionClassif[20]> 0.031000000
#>     resampling_id iteration              prediction  classif.ce

# OpenML task collection
task_collection = OMLCollection$new(id = 99)
# using sugar
task_collection = oml_collection(id = 99)

task_collection$main_entity_type
#> [1] "task"
task_collection$tasks
#>         id          task                           data
#>  1:      3 <OMLTask[25]>                       kr-vs-kp
#>  2:      6 <OMLTask[25]>                         letter
#>  3:     11 <OMLTask[25]>                  balance-scale
#>  4:     12 <OMLTask[25]>                  mfeat-factors
#>  5:     14 <OMLTask[25]>                  mfeat-fourier
#>  6:     15 <OMLTask[25]>                       breast-w
#>  7:     16 <OMLTask[25]>                 mfeat-karhunen
#>  8:     18 <OMLTask[25]>            mfeat-morphological
#>  9:     22 <OMLTask[25]>                  mfeat-zernike
#> 10:     23 <OMLTask[25]>                            cmc
#> 11:     28 <OMLTask[25]>                      optdigits
#> 12:     29 <OMLTask[25]>                credit-approval
#> 13:     31 <OMLTask[25]>                       credit-g
#> 14:     32 <OMLTask[25]>                      pendigits
#> 15:     37 <OMLTask[25]>                       diabetes
#> 16:     43 <OMLTask[25]>                       spambase
#> 17:     45 <OMLTask[25]>                         splice
#> 18:     49 <OMLTask[25]>                    tic-tac-toe
#> 19:     53 <OMLTask[25]>                        vehicle
#> 20:    219 <OMLTask[25]>                    electricity
#> 21:   2074 <OMLTask[25]>                       satimage
#> 22:   2079 <OMLTask[25]>                     eucalyptus
#> 23:   3021 <OMLTask[25]>                           sick
#> 24:   3022 <OMLTask[25]>                          vowel
#> 25:   3481 <OMLTask[25]>                         isolet
#> 26:   3549 <OMLTask[25]>         analcatdata_authorship
#> 27:   3560 <OMLTask[25]>               analcatdata_dmft
#> 28:   3573 <OMLTask[25]>                      mnist_784
#> 29:   3902 <OMLTask[25]>                            pc4
#> 30:   3903 <OMLTask[25]>                            pc3
#> 31:   3904 <OMLTask[25]>                            jm1
#> 32:   3913 <OMLTask[25]>                            kc2
#> 33:   3917 <OMLTask[25]>                            kc1
#> 34:   3918 <OMLTask[25]>                            pc1
#> 35:   7592 <OMLTask[25]>                          adult
#> 36:   9910 <OMLTask[25]>                    Bioresponse
#> 37:   9946 <OMLTask[25]>                           wdbc
#> 38:   9952 <OMLTask[25]>                        phoneme
#> 39:   9957 <OMLTask[25]>                    qsar-biodeg
#> 40:   9960 <OMLTask[25]>          wall-robot-navigation
#> 41:   9964 <OMLTask[25]>                        semeion
#> 42:   9971 <OMLTask[25]>                           ilpd
#> 43:   9976 <OMLTask[25]>                        madelon
#> 44:   9977 <OMLTask[25]>                          nomao
#> 45:   9978 <OMLTask[25]>                ozone-level-8hr
#> 46:   9981 <OMLTask[25]>                         cnae-9
#> 47:   9985 <OMLTask[25]>    first-order-theorem-proving
#> 48:  10093 <OMLTask[25]>        banknote-authentication
#> 49:  10101 <OMLTask[25]> blood-transfusion-service[...]
#> 50:  14952 <OMLTask[25]>               PhishingWebsites
#> 51:  14954 <OMLTask[25]>                 cylinder-bands
#> 52:  14965 <OMLTask[25]>                 bank-marketing
#> 53:  14969 <OMLTask[25]> GesturePhaseSegmentationP[...]
#> 54:  14970 <OMLTask[25]>                            har
#> 55: 125920 <OMLTask[25]>                  dresses-sales
#> 56: 125922 <OMLTask[25]>                        texture
#> 57: 146195 <OMLTask[25]>                      connect-4
#> 58: 146800 <OMLTask[25]>                    MiceProtein
#> 59: 146817 <OMLTask[25]>             steel-plates-fault
#> 60: 146819 <OMLTask[25]> climate-model-simulation-[...]
#> 61: 146820 <OMLTask[25]>                           wilt
#> 62: 146821 <OMLTask[25]>                            car
#> 63: 146822 <OMLTask[25]>                        segment
#> 64: 146824 <OMLTask[25]>                    mfeat-pixel
#> 65: 146825 <OMLTask[25]>                  Fashion-MNIST
#> 66: 167119 <OMLTask[25]> jungle_chess_2pcs_raw_end[...]
#> 67: 167120 <OMLTask[25]>                    numerai28.6
#> 68: 167121 <OMLTask[25]>               Devnagari-Script
#> 69: 167124 <OMLTask[25]>                       CIFAR_10
#> 70: 167125 <OMLTask[25]>        Internet-Advertisements
#> 71: 167140 <OMLTask[25]>                            dna
#> 72: 167141 <OMLTask[25]>                          churn
#>         id          task                           data
#>                     task_type                    target  nrow ncol missing
#>  1: Supervised Classification                     class  3196   37       0
#>  2: Supervised Classification                     class 20000   17       0
#>  3: Supervised Classification                     class   625    5       0
#>  4: Supervised Classification                     class  2000  217       0
#>  5: Supervised Classification                     class  2000   77       0
#>  6: Supervised Classification                     Class   699   10      16
#>  7: Supervised Classification                     class  2000   65       0
#>  8: Supervised Classification                     class  2000    7       0
#>  9: Supervised Classification                     class  2000   48       0
#> 10: Supervised Classification Contraceptive_method_used  1473   10       0
#> 11: Supervised Classification                     class  5620   65       0
#> 12: Supervised Classification                     class   690   16      67
#> 13: Supervised Classification                     class  1000   21       0
#> 14: Supervised Classification                     class 10992   17       0
#> 15: Supervised Classification                     class   768    9       0
#> 16: Supervised Classification                     class  4601   58       0
#> 17: Supervised Classification                     Class  3190   61       0
#> 18: Supervised Classification                     Class   958   10       0
#> 19: Supervised Classification                     Class   846   19       0
#> 20: Supervised Classification                     class 45312    9       0
#> 21: Supervised Classification                     class  6430   37       0
#> 22: Supervised Classification                   Utility   736   20     448
#> 23: Supervised Classification                     Class  3772   30    6064
#> 24: Supervised Classification                     Class   990   13       0
#> 25: Supervised Classification                     class  7797  618       0
#> 26: Supervised Classification                    Author   841   71       0
#> 27: Supervised Classification                Prevention   797    5       0
#> 28: Supervised Classification                     class 70000  785       0
#> 29: Supervised Classification                         c  1458   38       0
#> 30: Supervised Classification                         c  1563   38       0
#> 31: Supervised Classification                   defects 10885   22      25
#> 32: Supervised Classification                  problems   522   22       0
#> 33: Supervised Classification                   defects  2109   22       0
#> 34: Supervised Classification                   defects  1109   22       0
#> 35: Supervised Classification                     class 48842   15    6465
#> 36: Supervised Classification                    target  3751 1777       0
#> 37: Supervised Classification                     Class   569   31       0
#> 38: Supervised Classification                     Class  5404    6       0
#> 39: Supervised Classification                     Class  1055   42       0
#> 40: Supervised Classification                     Class  5456   25       0
#> 41: Supervised Classification                     Class  1593  257       0
#> 42: Supervised Classification                     Class   583   11       0
#> 43: Supervised Classification                     Class  2600  501       0
#> 44: Supervised Classification                     Class 34465  119       0
#> 45: Supervised Classification                     Class  2534   73       0
#> 46: Supervised Classification                     Class  1080  857       0
#> 47: Supervised Classification                     Class  6118   52       0
#> 48: Supervised Classification                     Class  1372    5       0
#> 49: Supervised Classification                     Class   748    5       0
#> 50: Supervised Classification                    Result 11055   31       0
#> 51: Supervised Classification                 band_type   540   40     999
#> 52: Supervised Classification                     Class 45211   17       0
#> 53: Supervised Classification                     Phase  9873   33       0
#> 54: Supervised Classification                     Class 10299  562       0
#> 55: Supervised Classification                     Class   500   13     835
#> 56: Supervised Classification                     Class  5500   41       0
#> 57: Supervised Classification                     class 67557   43       0
#> 58: Supervised Classification                     class  1080   82    1396
#> 59: Supervised Classification                    target  1941   28       0
#> 60: Supervised Classification                   outcome   540   21       0
#> 61: Supervised Classification                     class  4839    6       0
#> 62: Supervised Classification                     class  1728    7       0
#> 63: Supervised Classification                     class  2310   20       0
#> 64: Supervised Classification                     class  2000  241       0
#> 65: Supervised Classification                     class 70000  785       0
#> 66: Supervised Classification                     class 44819    7       0
#> 67: Supervised Classification              attribute_21 96320   22       0
#> 68: Supervised Classification                 character 92000 1025       0
#> 69: Supervised Classification                     class 60000 3073       0
#> 70: Supervised Classification                     class  3279 1559       0
#> 71: Supervised Classification                     class  3186  181       0
#> 72: Supervised Classification                     class  5000   21       0
#>                     task_type                    target  nrow ncol missing
#>     numeric symbolic binary     task_splits
#>  1:       0       37     35 crossvalidation
#>  2:      16        1      0 crossvalidation
#>  3:       4        1      0 crossvalidation
#>  4:     216        1      0 crossvalidation
#>  5:      76        1      0 crossvalidation
#>  6:       9        1      1 crossvalidation
#>  7:      64        1      0 crossvalidation
#>  8:       6        1      0 crossvalidation
#>  9:      47        1      0 crossvalidation
#> 10:       2        8      3 crossvalidation
#> 11:      64        1      0 crossvalidation
#> 12:       6       10      5 crossvalidation
#> 13:       7       14      3 crossvalidation
#> 14:      16        1      0 crossvalidation
#> 15:       8        1      1 crossvalidation
#> 16:      57        1      1 crossvalidation
#> 17:       0       61      0 crossvalidation
#> 18:       0       10      1 crossvalidation
#> 19:      18        1      0 crossvalidation
#> 20:       7        2      1 crossvalidation
#> 21:      36        1      0 crossvalidation
#> 22:      14        6      0 crossvalidation
#> 23:       7       23     21 crossvalidation
#> 24:      10        3      1 crossvalidation
#> 25:     617        1      0 crossvalidation
#> 26:      70        1      0 crossvalidation
#> 27:       0        5      1 crossvalidation
#> 28:     784        1      0 crossvalidation
#> 29:      37        1      1 crossvalidation
#> 30:      37        1      1 crossvalidation
#> 31:      21        1      1 crossvalidation
#> 32:      21        1      1 crossvalidation
#> 33:      21        1      1 crossvalidation
#> 34:      21        1      1 crossvalidation
#> 35:       6        9      2 crossvalidation
#> 36:    1776        1      1 crossvalidation
#> 37:      30        1      1 crossvalidation
#> 38:       5        1      1 crossvalidation
#> 39:      41        1      1 crossvalidation
#> 40:      24        1      0 crossvalidation
#> 41:     256        1      0 crossvalidation
#> 42:       9        2      2 crossvalidation
#> 43:     500        1      1 crossvalidation
#> 44:      89       30      3 crossvalidation
#> 45:      72        1      1 crossvalidation
#> 46:     856        1      0 crossvalidation
#> 47:      51        1      0 crossvalidation
#> 48:       4        1      1 crossvalidation
#> 49:       4        1      1 crossvalidation
#> 50:       0       31     23 crossvalidation
#> 51:      18       22      4 crossvalidation
#> 52:       7       10      4 crossvalidation
#> 53:      32        1      0 crossvalidation
#> 54:     561        1      0 crossvalidation
#> 55:       1       12      1 crossvalidation
#> 56:      40        1      0 crossvalidation
#> 57:       0       43      0 crossvalidation
#> 58:      77        5      3 crossvalidation
#> 59:      27        1      0 crossvalidation
#> 60:      20        1      1 crossvalidation
#> 61:       5        1      1 crossvalidation
#> 62:       0        7      0 crossvalidation
#> 63:      19        1      0 crossvalidation
#> 64:     240        1      0 crossvalidation
#> 65:     784        1      0 crossvalidation
#> 66:       6        1      0 crossvalidation
#> 67:      21        1      1 crossvalidation
#> 68:    1024        1      0 crossvalidation
#> 69:    3072        1      0 crossvalidation
#> 70:       3     1556   1556 crossvalidation
#> 71:       0      181    180 crossvalidation
#> 72:      16        5      3 crossvalidation
#>     numeric symbolic binary     task_splits
task_collection$data
#>        id          data                                   name  nrow ncol
#>  1:     3 <OMLData[25]>                               kr-vs-kp  3196   37
#>  2:     6 <OMLData[25]>                                 letter 20000   17
#>  3:    11 <OMLData[25]>                          balance-scale   625    5
#>  4:    12 <OMLData[25]>                          mfeat-factors  2000  217
#>  5:    14 <OMLData[25]>                          mfeat-fourier  2000   77
#>  6:    15 <OMLData[25]>                               breast-w   699   10
#>  7:    16 <OMLData[25]>                         mfeat-karhunen  2000   65
#>  8:    18 <OMLData[25]>                    mfeat-morphological  2000    7
#>  9:    22 <OMLData[25]>                          mfeat-zernike  2000   48
#> 10:    23 <OMLData[25]>                                    cmc  1473   10
#> 11:    28 <OMLData[25]>                              optdigits  5620   65
#> 12:    29 <OMLData[25]>                        credit-approval   690   16
#> 13:    31 <OMLData[25]>                               credit-g  1000   21
#> 14:    32 <OMLData[25]>                              pendigits 10992   17
#> 15:    37 <OMLData[25]>                               diabetes   768    9
#> 16:    38 <OMLData[25]>                                   sick  3772   30
#> 17:    44 <OMLData[25]>                               spambase  4601   58
#> 18:    46 <OMLData[25]>                                 splice  3190   61
#> 19:    50 <OMLData[25]>                            tic-tac-toe   958   10
#> 20:    54 <OMLData[25]>                                vehicle   846   19
#> 21:   151 <OMLData[25]>                            electricity 45312    9
#> 22:   182 <OMLData[25]>                               satimage  6430   37
#> 23:   188 <OMLData[25]>                             eucalyptus   736   20
#> 24:   300 <OMLData[25]>                                 isolet  7797  618
#> 25:   307 <OMLData[25]>                                  vowel   990   13
#> 26:   458 <OMLData[25]>                 analcatdata_authorship   841   71
#> 27:   469 <OMLData[25]>                       analcatdata_dmft   797    5
#> 28:   554 <OMLData[25]>                              mnist_784 70000  785
#> 29:  1049 <OMLData[25]>                                    pc4  1458   38
#> 30:  1050 <OMLData[25]>                                    pc3  1563   38
#> 31:  1053 <OMLData[25]>                                    jm1 10885   22
#> 32:  1063 <OMLData[25]>                                    kc2   522   22
#> 33:  1067 <OMLData[25]>                                    kc1  2109   22
#> 34:  1068 <OMLData[25]>                                    pc1  1109   22
#> 35:  1461 <OMLData[25]>                         bank-marketing 45211   17
#> 36:  1462 <OMLData[25]>                banknote-authentication  1372    5
#> 37:  1464 <OMLData[25]>       blood-transfusion-service-center   748    5
#> 38:  1468 <OMLData[25]>                                 cnae-9  1080  857
#> 39:  1475 <OMLData[25]>            first-order-theorem-proving  6118   52
#> 40:  1478 <OMLData[25]>                                    har 10299  562
#> 41:  1480 <OMLData[25]>                                   ilpd   583   11
#> 42:  1485 <OMLData[25]>                                madelon  2600  501
#> 43:  1486 <OMLData[25]>                                  nomao 34465  119
#> 44:  1487 <OMLData[25]>                        ozone-level-8hr  2534   73
#> 45:  1489 <OMLData[25]>                                phoneme  5404    6
#> 46:  1494 <OMLData[25]>                            qsar-biodeg  1055   42
#> 47:  1497 <OMLData[25]>                  wall-robot-navigation  5456   25
#> 48:  1501 <OMLData[25]>                                semeion  1593  257
#> 49:  1510 <OMLData[25]>                                   wdbc   569   31
#> 50:  1590 <OMLData[25]>                                  adult 48842   15
#> 51:  4134 <OMLData[25]>                            Bioresponse  3751 1777
#> 52:  4534 <OMLData[25]>                       PhishingWebsites 11055   31
#> 53:  4538 <OMLData[25]>      GesturePhaseSegmentationProcessed  9873   33
#> 54:  6332 <OMLData[25]>                         cylinder-bands   540   40
#> 55: 23381 <OMLData[25]>                          dresses-sales   500   13
#> 56: 23517 <OMLData[25]>                            numerai28.6 96320   22
#> 57: 40499 <OMLData[25]>                                texture  5500   41
#> 58: 40668 <OMLData[25]>                              connect-4 67557   43
#> 59: 40670 <OMLData[25]>                                    dna  3186  181
#> 60: 40701 <OMLData[25]>                                  churn  5000   21
#> 61: 40923 <OMLData[25]>                       Devnagari-Script 92000 1025
#> 62: 40927 <OMLData[25]>                               CIFAR_10 60000 3073
#> 63: 40966 <OMLData[25]>                            MiceProtein  1080   82
#> 64: 40975 <OMLData[25]>                                    car  1728    7
#> 65: 40978 <OMLData[25]>                Internet-Advertisements  3279 1559
#> 66: 40979 <OMLData[25]>                            mfeat-pixel  2000  241
#> 67: 40982 <OMLData[25]>                     steel-plates-fault  1941   28
#> 68: 40983 <OMLData[25]>                                   wilt  4839    6
#> 69: 40984 <OMLData[25]>                                segment  2310   20
#> 70: 40994 <OMLData[25]>       climate-model-simulation-crashes   540   21
#> 71: 40996 <OMLData[25]>                          Fashion-MNIST 70000  785
#> 72: 41027 <OMLData[25]> jungle_chess_2pcs_raw_endgame_complete 44819    7
#>        id          data                                   name  nrow ncol
#>     missing numeric symbolic binary
#>  1:       0       0       37     35
#>  2:       0      16        1      0
#>  3:       0       4        1      0
#>  4:       0     216        1      0
#>  5:       0      76        1      0
#>  6:      16       9        1      1
#>  7:       0      64        1      0
#>  8:       0       6        1      0
#>  9:       0      47        1      0
#> 10:       0       2        8      3
#> 11:       0      64        1      0
#> 12:      67       6       10      5
#> 13:       0       7       14      3
#> 14:       0      16        1      0
#> 15:       0       8        1      1
#> 16:    6064       7       23     21
#> 17:       0      57        1      1
#> 18:       0       0       61      0
#> 19:       0       0       10      1
#> 20:       0      18        1      0
#> 21:       0       7        2      1
#> 22:       0      36        1      0
#> 23:     448      14        6      0
#> 24:       0     617        1      0
#> 25:       0      10        3      1
#> 26:       0      70        1      0
#> 27:       0       0        5      1
#> 28:       0     784        1      0
#> 29:       0      37        1      1
#> 30:       0      37        1      1
#> 31:      25      21        1      1
#> 32:       0      21        1      1
#> 33:       0      21        1      1
#> 34:       0      21        1      1
#> 35:       0       7       10      4
#> 36:       0       4        1      1
#> 37:       0       4        1      1
#> 38:       0     856        1      0
#> 39:       0      51        1      0
#> 40:       0     561        1      0
#> 41:       0       9        2      2
#> 42:       0     500        1      1
#> 43:       0      89       30      3
#> 44:       0      72        1      1
#> 45:       0       5        1      1
#> 46:       0      41        1      1
#> 47:       0      24        1      0
#> 48:       0     256        1      0
#> 49:       0      30        1      1
#> 50:    6465       6        9      2
#> 51:       0    1776        1      1
#> 52:       0       0       31     23
#> 53:       0      32        1      0
#> 54:     999      18       22      4
#> 55:     835       1       12      1
#> 56:       0      21        1      1
#> 57:       0      40        1      0
#> 58:       0       0       43      0
#> 59:       0       0      181    180
#> 60:       0      16        5      3
#> 61:       0    1024        1      0
#> 62:       0    3072        1      0
#> 63:    1396      77        5      3
#> 64:       0       0        7      0
#> 65:       0       3     1556   1556
#> 66:       0     240        1      0
#> 67:       0      27        1      0
#> 68:       0       5        1      1
#> 69:       0      19        1      0
#> 70:       0      20        1      1
#> 71:       0     784        1      0
#> 72:       0       6        1      0
#>     missing numeric symbolic binary

# mlr3 conversion
tasks = as_tasks(task_collection)
resamplings = as_resamplings(task_collection)
# }