Interface to OpenML Flows

This is the class for flows served on OpenML. Flows represent machine learning algorithms. This object can also be constructed using the sugar function oml_flow().

mlr3 Integration

Obtain a mlr3::Learner using mlr3::as_learner().

References

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi:10.1145/2641190.2641198 .

Super class

mlr3oml::OMLObject -> OMLFlow

Active bindings

parameter

(data.table)
The parameters of the flow.

dependencies

(character())
The dependencies of the flow.

tags

(character())
Returns all tags of the object.

Methods

Public methods

OMLFlow$new()

OMLFlow$print()

OMLFlow$clone()

Inherited methods

mlr3oml::OMLObject$help()

Method new()

Creates a new instance of this R6 class.

Usage

OMLFlow$new(id, cache = cache_default(), test_server = test_server_default())

Arguments

id

(integer(1))
OpenML id for the object.

cache

(logical(1) | character(1))
See field cache for an explanation of possible values. Defaults to value of option "mlr3oml.cache", or FALSE if not set.

test_server

(character(1))
Whether to use the OpenML test server (https://test.openml.org/) or public server (https://www.openml.org/). Defaults to value of option "mlr3oml.test_server", or FALSE if not set.

Method print()

Prints the object.

Usage

OMLFlow$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

OMLFlow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library("mlr3")
# mlr3 flow:
flow = OMLFlow$new(id = 19103)
# using sugar
flow = oml_flow(id = 19103)
learner = as_learner(flow, "classif")
# python flow
python_flow = OMLFlow$new(19090)
# conversion to pseudo Learner
plearner = as_learner(python_flow, "classif")
# }