List Tasks from OpenML

This function allows to find tasks on https://openml.org/t using some simple filter criteria.

Note that only a subset of filters is exposed here. For a more feature-complete package, see OpenML.

list_oml_tasks(
  task_id = NULL,
  number_instances = NULL,
  number_features = NULL,
  number_classes = NULL,
  number_missing_values = NULL,
  tag = NULL,
  limit = 5000L,
  ...
)

Arguments

task_id
(integer())
Vector of task ids to restrict to.

number_instances
(integer())
Filter for number of instances.

number_features
(integer())
Filter for number of features.

number_classes
(integer())
Filter for number of labels of the target (only classification tasks).

number_missing_values
(integer())
Filter for number of missing values.

tag
(character())
Filter for specific tag. You can provide multiple tags as character vector.

limit
(integer())
Limit the results to limit records. Default is 5000.

...
(any)
Additional filters as named arguments.

Value

(data.table()) of results, or a Null data.table if no task matches the criteria.

Details

Filter values can be provided as single atomic values (typically integer or character). Provide a numeric vector of length 2 (c(l, u)) to find matches in the range \([l, u]\).

References

Casalicchio G, Bossek J, Lang M, Kirchhoff D, Kerschke P, Hofner B, Seibold H, Vanschoren J, Bischl B (2017). “OpenML: An R Package to Connect to the Machine Learning Platform OpenML.” Computational Statistics, 1--15. doi: 10.1007/s00180-017-0742-2 .

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi: 10.1145/2641190.2641198 .

Examples

# \donttest{
list_oml_tasks(number_instances = 150, number_features = c(1, 10))

#>      task_id                             task_type data_id
#>   1:      59             Supervised Classification      61
#>   2:     118                        Learning Curve      61
#>   3:     289             Supervised Classification      61
#>   4:    1758                        Learning Curve      61
#>   5:    1823             Supervised Classification      61
#>  ---                                                      
#> 483:  360562                        Learning Curve   42186
#> 484:  360570                        Learning Curve   42535
#> 485:  360584                            Clustering      61
#> 486:  360888 Supervised Data Stream Classification   42186
#> 487:  360896 Supervised Data Stream Classification   42535
#>                                  name status MajorityClassSize
#>   1:                             iris active                50
#>   2:                             iris active                50
#>   3:                             iris active                50
#>   4:                             iris active                50
#>   5:                             iris active                50
#>  ---                                                          
#> 483:                  JuanFeldmanIris active                50
#> 484: TEST10e627dcde-UploadTestWithURL active                NA
#> 485:                             iris active                50
#> 486:                  JuanFeldmanIris active                50
#> 487: TEST10e627dcde-UploadTestWithURL active                NA
#>      MaxNominalAttDistinctValues MinorityClassSize NumberOfClasses
#>   1:                           3                50               3
#>   2:                           3                50               3
#>   3:                           3                50               3
#>   4:                           3                50               3
#>   5:                           3                50               3
#>  ---                                                              
#> 483:                           3                50               3
#> 484:                          NA                NA              NA
#> 485:                           3                50               3
#> 486:                           3                50               3
#> 487:                          NA                NA              NA
#>      NumberOfFeatures NumberOfInstances NumberOfInstancesWithMissingValues
#>   1:                5               150                                  0
#>   2:                5               150                                  0
#>   3:                5               150                                  0
#>   4:                5               150                                  0
#>   5:                5               150                                  0
#>  ---                                                                      
#> 483:                5               150                                  0
#> 484:                5               150                                  0
#> 485:                5               150                                  0
#> 486:                5               150                                  0
#> 487:                5               150                                  0
#>      NumberOfMissingValues NumberOfNumericFeatures NumberOfSymbolicFeatures
#>   1:                     0                       4                        1
#>   2:                     0                       4                        1
#>   3:                     0                       4                        1
#>   4:                     0                       4                        1
#>   5:                     0                       4                        1
#>  ---                                                                       
#> 483:                     0                       4                        1
#> 484:                     0                       4                        1
#> 485:                     0                       4                        1
#> 486:                     0                       4                        1
#> 487:                     0                       4                        1

# }