List Tasks from OpenML

This function allows to find tasks on https://openml.org/t using some simple filter criteria.

Note that only a subset of filters is exposed here. For a more feature-complete package, see OpenML.

list_oml_tasks(
  task_id = NULL,
  number_instances = NULL,
  number_features = NULL,
  number_classes = NULL,
  number_missing_values = NULL,
  tag = NULL,
  limit = 5000L,
  ...
)

Arguments

task_id
(integer())
Vector of task ids to restrict to.

number_instances
(integer())
Filter for number of instances.

number_features
(integer())
Filter for number of features.

number_classes
(integer())
Filter for number of labels of the target (only classification tasks).

number_missing_values
(integer())
Filter for number of missing values.

tag
(character())
Filter for specific tag. You can provide multiple tags as character vector.

limit
(integer())
Limit the results to limit records. Default is 5000.

...
(any)
Additional filters as named arguments.

Value

(data.table()) of results, or NULL if no data set matches the criteria.

Details

Filter values can be provided as single atomic values (typically integer or character). Provide a numeric vector of length 2 (c(l, u)) to find matches in the range \([l, u]\).

References

Casalicchio G, Bossek J, Lang M, Kirchhoff D, Kerschke P, Hofner B, Seibold H, Vanschoren J, Bischl B (2017). “OpenML: An R Package to Connect to the Machine Learning Platform OpenML.” Computational Statistics, 1--15. doi: 10.1007/s00180-017-0742-2 . Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi: 10.1145/2641190.2641198 .

Examples

# \donttest{
list_oml_tasks(number_instances = 150, number_features = c(1, 10))

#>      task_id                 task_type data_id                             name
#>   1:      59 Supervised Classification      61                             iris
#>   2:     118            Learning Curve      61                             iris
#>   3:     289 Supervised Classification      61                             iris
#>   4:    1758            Learning Curve      61                             iris
#>   5:    1823 Supervised Classification      61                             iris
#>  ---                                                                           
#> 367:  317211                Clustering   42097                             iris
#> 368:  317212                Clustering   42098                             iris
#> 369:  317279                Clustering   42186                  JuanFeldmanIris
#> 370:  317326                Clustering   42261                     iris-example
#> 371:  317501                Clustering   42535 TEST10e627dcde-UploadTestWithURL
#>      status MajorityClassSize MaxNominalAttDistinctValues MinorityClassSize
#>   1: active                50                           3                50
#>   2: active                50                           3                50
#>   3: active                50                           3                50
#>   4: active                50                           3                50
#>   5: active                50                           3                50
#>  ---                                                                       
#> 367: active                NA                           3                NA
#> 368: active                50                           3                50
#> 369: active                50                           3                50
#> 370: active                50                          NA                50
#> 371: active                NA                          NA                NA
#>      NumberOfClasses NumberOfFeatures NumberOfInstances
#>   1:               3                5               150
#>   2:               3                5               150
#>   3:               3                5               150
#>   4:               3                5               150
#>   5:               3                5               150
#>  ---                                                   
#> 367:              NA                5               150
#> 368:               3                5               150
#> 369:               3                5               150
#> 370:               3                5               150
#> 371:              NA                5               150
#>      NumberOfInstancesWithMissingValues NumberOfMissingValues
#>   1:                                  0                     0
#>   2:                                  0                     0
#>   3:                                  0                     0
#>   4:                                  0                     0
#>   5:                                  0                     0
#>  ---                                                         
#> 367:                                  0                     0
#> 368:                                  0                     0
#> 369:                                  0                     0
#> 370:                                  0                     0
#> 371:                                  0                     0
#>      NumberOfNumericFeatures NumberOfSymbolicFeatures
#>   1:                       4                        1
#>   2:                       4                        1
#>   3:                       4                        1
#>   4:                       4                        1
#>   5:                       4                        1
#>  ---                                                 
#> 367:                       4                        1
#> 368:                       4                        1
#> 369:                       4                        1
#> 370:                       4                        1
#> 371:                       4                        1

# }