OpenML Collection

This is the class for collections (previously known as studies) served on https://www.openml.org. A collection can either be a task collection or run collection. This object can also be constructed using the sugar function oml_collection().

Run Collection

A run collection contains runs, flows, datasets and tasks. The primary object are the runs (main_entity_type is "run"). The the flows, datasets and tasks are those used in the runs.

Task Collection A task collection (main_entity_type = "task") contains tasks and datasets. The primary object are the tasks (main_entity_type is "task"). The datasets are those used in the tasks.

Note: All Benchmark Suites on OpenML are also collections.

Caching

The OpenML collection itself cannot be not cached, this is because it can be modified in-place on the server, e.g. by adding or removing tasks or runs. The construction argument cache therefore only controls wether caching is applied to the OpenML objects that are contained in the collection.

mlr3 Intergration

Obtain a list of mlr3::Tasks using mlr3::as_tasks.

Obtain a list of mlr3::Resamplings using mlr3::as_resamplings.

Obtain a list of mlr3::Learners using mlr3::as_learners (if main_entity_type is "run").

Obtain a mlr3::BenchmarkResult using mlr3::as_benchmark_result (if main_entity_type is "run").

References

Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014). “OpenML.” ACM SIGKDD Explorations Newsletter, 15(2), 49--60. doi:10.1145/2641190.2641198 .

Super class

mlr3oml::OMLObject -> OMLCollection

Active bindings

desc

(list())
Colllection description (meta information), downloaded and converted from the JSON API response.

parquet

(logical(1))
Whether to use parquet.

main_entity_type

(character(n))
The main entity type, either "run" or "task".

flow_ids

(integer(n))
An vector containing the flow ids of the collection.

data_ids

(integer(n))
An vector containing the data ids of the collection.

run_ids

(integer(n))
An vector containing the run ids of the collection.

task_ids

(integer(n))
An vector containing the task ids of the collection.

runs

(data.table()) A data.table summarizing the runs included in the collection. Returns NULL for Task Collections.

flows

(data.table()) A data.table summarizing the flows included in the collection. Returns NULL for Task Collections.

data

(data.table()) A data.table summarizing the datasets included in the collection.

tasks

(data.table()) A data.table summarizing the tasks included in the collection.

Methods

Public methods

OMLCollection$new()

OMLCollection$print()

OMLCollection$clone()

Inherited methods

mlr3oml::OMLObject$help()

Method new()

Creates a new instance of this R6 class.

Usage

OMLCollection$new(
  id,
  cache = cache_default(),
  parquet = parquet_default(),
  test_server = test_server_default()
)

Arguments

id

(integer(1))
OpenML id for the object.

cache

(logical(1) | character(1))
See field cache for an explanation of possible values. Defaults to value of option "mlr3oml.cache", or FALSE if not set. The collection itself is not cached, this is because it can be modified in-place on OpenML, e.g. by adding or removing tasks or runs. This parameter therefore only controls whether the contained elements are cached when loaded, e.g. when accessing the included tasks.

parquet

(logical(1))
Whether to use parquet instead of arff. If parquet is not available, it will fall back to arff. Defaults to value of option "mlr3oml.parquet" or FALSE if not set.

test_server

(character(1))
Whether to use the OpenML test server (https://test.openml.org/) or public server (https://www.openml.org/). Defaults to value of option "mlr3oml.test_server", or FALSE if not set.

Method print()

Prints the object.

Usage

OMLCollection$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

OMLCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
library("mlr3")
# OpenML Run collection:
run_collection = OMLCollection$new(id = 232)
# using sugar
run_collection = oml_collection(id = 232)
run_collection$main_entity_type
#> [1] "run"
run_collection$tasks
#>    id          task     data                 task_type target  nrow ncol
#> 1:  3 <OMLTask[25]> kr-vs-kp Supervised Classification  class  3196   37
#> 2:  6 <OMLTask[25]>   letter Supervised Classification  class 20000   17
#>    missing numeric symbolic binary     task_splits
#> 1:       0       0       37     35 crossvalidation
#> 2:       0      16        1      0 crossvalidation
run_collection$data
#>    id          data     name  nrow ncol missing numeric symbolic binary
#> 1:  3 <OMLData[25]> kr-vs-kp  3196   37       0       0       37     35
#> 2:  6 <OMLData[25]>   letter 20000   17       0      16        1      0
run_collection$flows
#>       id          flow                           name
#> 1: 17369 <OMLFlow[16]> sklearn.model_selection._[...]
#> 2: 17374 <OMLFlow[16]> sklearn.model_selection._[...]
run_collection$runs
#>          id          run                 task_type     data
#> 1: 10417460 <OMLRun[25]> Supervised Classification kr-vs-kp
#> 2: 10417461 <OMLRun[25]> Supervised Classification kr-vs-kp
#> 3: 10417462 <OMLRun[25]> Supervised Classification   letter
#> 4: 10417463 <OMLRun[25]> Supervised Classification   letter
#>                              flow     task_splits
#> 1: sklearn.model_selection._[...] crossvalidation
#> 2: sklearn.model_selection._[...] crossvalidation
#> 3: sklearn.model_selection._[...] crossvalidation
#> 4: sklearn.model_selection._[...] crossvalidation

# mlr3 conversion:
tasks = as_tasks(run_collection)
resamplings = as_resamplings(run_collection)
learners = as_learners(run_collection, "classif")

bmr = as_benchmark_result(run_collection)
bmr$score(msr("classif.ce"))
#>                                    uhash nr              task  task_id
#>  1: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  2: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  3: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  4: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  5: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  6: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  7: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  8: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#>  9: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#> 10: b27c7c2a-b269-4176-8406-b0d9fb01715e  1 <TaskClassif[50]> kr-vs-kp
#> 11: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 12: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 13: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 14: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 15: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 16: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 17: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 18: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 19: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 20: 790fa18e-8c6c-44d6-bf71-e85f47c0ec9d  2 <TaskClassif[50]> kr-vs-kp
#> 21: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 22: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 23: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 24: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 25: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 26: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 27: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 28: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 29: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 30: 7e55dde4-9f7e-4322-9b97-a93eed3c21d4  3 <TaskClassif[50]>   letter
#> 31: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 32: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 33: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 34: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 35: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 36: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 37: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 38: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 39: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#> 40: d3e3e71a-8bbf-4af8-b119-a004f6d2ec9f  4 <TaskClassif[50]>   letter
#>                                    uhash nr              task  task_id
#>                          learner learner_id             resampling
#>  1: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  2: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  3: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  4: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  5: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  6: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  7: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  8: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#>  9: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 10: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 11: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 12: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 13: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 14: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 15: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 16: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 17: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 18: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 19: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 20: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 21: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 22: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 23: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 24: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 25: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 26: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 27: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 28: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 29: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 30: <LearnerClassifOML17374[36]>  oml.17374 <ResamplingCustom[20]>
#> 31: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 32: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 33: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 34: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 35: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 36: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 37: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 38: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 39: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#> 40: <LearnerClassifOML17369[36]>  oml.17369 <ResamplingCustom[20]>
#>                          learner learner_id             resampling
#>     resampling_id iteration              prediction  classif.ce
#>  1:        custom         1 <PredictionClassif[20]> 0.003134796
#>  2:        custom         2 <PredictionClassif[20]> 0.009375000
#>  3:        custom         3 <PredictionClassif[20]> 0.000000000
#>  4:        custom         4 <PredictionClassif[20]> 0.003125000
#>  5:        custom         5 <PredictionClassif[20]> 0.006250000
#>  6:        custom         6 <PredictionClassif[20]> 0.000000000
#>  7:        custom         7 <PredictionClassif[20]> 0.003134796
#>  8:        custom         8 <PredictionClassif[20]> 0.006269592
#>  9:        custom         9 <PredictionClassif[20]> 0.003134796
#> 10:        custom        10 <PredictionClassif[20]> 0.006269592
#> 11:        custom         1 <PredictionClassif[20]> 0.006269592
#> 12:        custom         2 <PredictionClassif[20]> 0.009375000
#> 13:        custom         3 <PredictionClassif[20]> 0.003125000
#> 14:        custom         4 <PredictionClassif[20]> 0.003125000
#> 15:        custom         5 <PredictionClassif[20]> 0.012500000
#> 16:        custom         6 <PredictionClassif[20]> 0.000000000
#> 17:        custom         7 <PredictionClassif[20]> 0.003134796
#> 18:        custom         8 <PredictionClassif[20]> 0.009404389
#> 19:        custom         9 <PredictionClassif[20]> 0.006269592
#> 20:        custom        10 <PredictionClassif[20]> 0.009404389
#> 21:        custom         1 <PredictionClassif[20]> 0.047523762
#> 22:        custom         2 <PredictionClassif[20]> 0.039500000
#> 23:        custom         3 <PredictionClassif[20]> 0.043500000
#> 24:        custom         4 <PredictionClassif[20]> 0.041000000
#> 25:        custom         5 <PredictionClassif[20]> 0.030500000
#> 26:        custom         6 <PredictionClassif[20]> 0.044000000
#> 27:        custom         7 <PredictionClassif[20]> 0.045000000
#> 28:        custom         8 <PredictionClassif[20]> 0.036000000
#> 29:        custom         9 <PredictionClassif[20]> 0.039000000
#> 30:        custom        10 <PredictionClassif[20]> 0.031000000
#> 31:        custom         1 <PredictionClassif[20]> 0.047523762
#> 32:        custom         2 <PredictionClassif[20]> 0.039500000
#> 33:        custom         3 <PredictionClassif[20]> 0.043500000
#> 34:        custom         4 <PredictionClassif[20]> 0.041000000
#> 35:        custom         5 <PredictionClassif[20]> 0.030500000
#> 36:        custom         6 <PredictionClassif[20]> 0.044000000
#> 37:        custom         7 <PredictionClassif[20]> 0.045000000
#> 38:        custom         8 <PredictionClassif[20]> 0.036000000
#> 39:        custom         9 <PredictionClassif[20]> 0.039000000
#> 40:        custom        10 <PredictionClassif[20]> 0.031000000
#>     resampling_id iteration              prediction  classif.ce

# OpenML task collection
task_collection = OMLCollection$new(id = 258)
# using sugar
task_collection = oml_collection(id = 258)

task_collection$main_entity_type
#> [1] "task"
task_collection$tasks
#>         id          task                    data                 task_type
#>  1:     20 <OMLTask[25]>             mfeat-pixel Supervised Classification
#>  2:     28 <OMLTask[25]>               optdigits Supervised Classification
#>  3:     29 <OMLTask[25]>         credit-approval Supervised Classification
#>  4:     32 <OMLTask[25]>               pendigits Supervised Classification
#>  5:     37 <OMLTask[25]>                diabetes Supervised Classification
#>  6:     49 <OMLTask[25]>             tic-tac-toe Supervised Classification
#>  7:   3549 <OMLTask[25]>  analcatdata_authorship Supervised Classification
#>  8:   3560 <OMLTask[25]>        analcatdata_dmft Supervised Classification
#>  9:  10093 <OMLTask[25]> banknote-authentication Supervised Classification
#> 10:  14970 <OMLTask[25]>                     har Supervised Classification
#> 11: 125922 <OMLTask[25]>                 texture Supervised Classification
#> 12: 146227 <OMLTask[25]>                   churn Supervised Classification
#>         target  nrow ncol missing numeric symbolic binary     task_splits
#>  1:      class  2000  241       0       0      241      0 crossvalidation
#>  2:      class  5620   65       0      64        1      0 crossvalidation
#>  3:      class   690   16      67       6       10      5 crossvalidation
#>  4:      class 10992   17       0      16        1      0 crossvalidation
#>  5:      class   768    9       0       8        1      1 crossvalidation
#>  6:      Class   958   10       0       0       10      1 crossvalidation
#>  7:     Author   841   71       0      70        1      0 crossvalidation
#>  8: Prevention   797    5       0       0        5      1 crossvalidation
#>  9:      Class  1372    5       0       4        1      1 crossvalidation
#> 10:      Class 10299  562       0     561        1      0 crossvalidation
#> 11:      Class  5500   41       0      40        1      0 crossvalidation
#> 12:      class  5000   21       0      16        5      3 crossvalidation
task_collection$data
#>        id          data                    name  nrow ncol missing numeric
#>  1:    20 <OMLData[25]>             mfeat-pixel  2000  241       0       0
#>  2:    28 <OMLData[25]>               optdigits  5620   65       0      64
#>  3:    29 <OMLData[25]>         credit-approval   690   16      67       6
#>  4:    32 <OMLData[25]>               pendigits 10992   17       0      16
#>  5:    37 <OMLData[25]>                diabetes   768    9       0       8
#>  6:    50 <OMLData[25]>             tic-tac-toe   958   10       0       0
#>  7:   458 <OMLData[25]>  analcatdata_authorship   841   71       0      70
#>  8:   469 <OMLData[25]>        analcatdata_dmft   797    5       0       0
#>  9:  1462 <OMLData[25]> banknote-authentication  1372    5       0       4
#> 10:  1478 <OMLData[25]>                     har 10299  562       0     561
#> 11: 40499 <OMLData[25]>                 texture  5500   41       0      40
#> 12: 40701 <OMLData[25]>                   churn  5000   21       0      16
#>     symbolic binary
#>  1:      241      0
#>  2:        1      0
#>  3:       10      5
#>  4:        1      0
#>  5:        1      1
#>  6:       10      1
#>  7:        1      0
#>  8:        5      1
#>  9:        1      1
#> 10:        1      0
#> 11:        1      0
#> 12:        5      3

# mlr3 conversion
tasks = as_tasks(task_collection)
resamplings = as_resamplings(task_collection)
# }